editoast: move post paced trains function to train schedule set view
